requirements.txt
ansi2html==1.9.1
asgiref==3.7.2
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bokeh==3.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
ciecplib==0.8.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==5.0
djangorestframework==3.14.0
environs==10.0.0
Flask==3.0.0
html2text==2020.1.16
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
linkify-it-py==2.0.2
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.4
panel==1.3.4
param==2.0.1
Pillow==10.1.0
plotly==5.18.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyviz_comms==3.0.0
PyYAML==6.0.1
requests==2.31.0
requests-ecp==0.3.1
retrying==1.3.4
setuptools==69.0.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tenacity==8.2.3
tomli==2.0.1
tomli_w==1.0.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.42.0
xyzservices==2023.10.1
zipp==3.17.0
